✨ 게시물 수정 다이얼로그 widget으로 변경
