--- src/feature/editPost/ui/index.ts
@@ -1,1 +1,0 @@
-export * from './EditPost';
